# pkg/parcacol: NormalizeWriteRawRequest returns flat slice of samples