This will get rid of the nested `series > profiles > samples`. We need to pass the labels and meta into each normalizedSample but in the end, don't need to loop more than once to ingest.

This makes distributing samples by hashing them a lot easier too.

```
name                              old time/op    new time/op    delta
ProfileColumnStoreWriteSeries-24     451ns ±17%    1430ns ±11%  +216.89%  (p=0.000 n=10+10)

name                              old alloc/op   new alloc/op   delta
ProfileColumnStoreWriteSeries-24      563B ± 0%     4568B ± 0%  +711.87%  (p=0.000 n=10+10)

name                              old allocs/op  new allocs/op  delta
ProfileColumnStoreWriteSeries-24      7.00 ± 0%      7.00 ± 0%      ~     (all equal)
```
The performance is quite a bit worse, but given we just came from a -99% decrease in #2202, I'd say the ergonomics are more important than the performance here.