@metalmatze 2022-12-01
Seems like the improvements I was hoping for over what we had on `main` don't really make a big difference downstream. So closing this.